requirements.txt
Flask==2.1.0
Flask-Cors==3.1.1
numpy==1.21.5
Pillow==8.4.0
tensorflow==2.7.0
